significantly improved ai. added easy and hard

+++ core/src/main/java/no/kash/gamedev/jag/game/gameobjects/players/damagehandlers/VanillaDamageHandler.java
@@ -16,7 +16,7 @@ public class VanillaDamageHandler implements DamageHandler {
 
 	@Override
 	public void onDamage(Bullet bullet) {
-		player.setHealth(player.getHealth() - bullet.getDamage());
+		player.setHealth((int)(player.getHealth() - bullet.getDamage()));
 		if (player.getHealth() <= 0) {
 			player.death(bullet.getShooter());
 		}
@@ -25,7 +25,7 @@ public class VanillaDamageHandler implements DamageHandler {
 	@Override
 	public void onDamage(Explosion explosion) {
 		float damage = 100 - Math.min(player.distanceTo(explosion), 100);
-		player.setHealth(player.getHealth() - damage);
+		player.setHealth((int)(player.getHealth() - damage));
 		if (player.getHealth() <= 0) {
 			player.death(explosion.getSourceGrenade().getThrower());
 		}
@@ -33,7 +33,7 @@ public class VanillaDamageHandler implements DamageHandler {
 
 	@Override
 	public void onDamage(Status status) {
-		player.setHealth(player.getHealth() - status.getDamage());
+		player.setHealth((int)(player.getHealth() - status.getDamage()));
 		if (player.getHealth() <= 0) {
 			player.death(status);
 		}
